requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
flask-swagger-ui==4.11.1
Flask-WTF==1.2.1
gunicorn==21.2.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mongoengine==0.27.0
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pymongo==4.6.1
pymssql==2.3.0
pytest==7.4.3
python-decouple==3.8
python-dotenv==1.0.0
Werkzeug==3.0.1
WTForms==3.1.1
